+++ src/holyc_lib.mli
@@ -31,8 +31,8 @@ module X86_64_expression = Backend.X86_64_expression
 module X86_64_program = Backend.X86_64_program
 module Native_execution = Runtime.Native_execution
 module Native_program_execution = Runtime.Native_program_execution
-module Native_expression = Driver.Native_expression
-module Native_program = Driver.Native_program
+module Native_expression = Hosted.Native_expression
+module Native_program = Hosted.Native_program
 
 module Ir_integer_globals : sig
   type t = Ir.Integer_globals.t
@@ -113,6 +113,7 @@ end
 
 module Ir_integer_interpreter = Ir.Integer_interpreter
 module Ir_runtime_call_context = Ir.Runtime_call_context
+module Native_parameter_defaults = Driver.Native_parameter_defaults
 module Ir_integer_program_lowering = Ir.Integer_program_lowering
 module Ir_global_initialization = Ir.Global_initialization